move some typedefs to a separate header
