+++ src/nanoarrow/nanoarrow.h
@@ -22,96 +22,11 @@
 #include <stdint.h>
 #include <stdlib.h>
 
+#include "typedefs_inline.h"
+
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-// Extra guard for versions of Arrow without the canonical guard
-#ifndef ARROW_FLAG_DICTIONARY_ORDERED
-
-#ifndef ARROW_C_DATA_INTERFACE
-#define ARROW_C_DATA_INTERFACE
-
-#define ARROW_FLAG_DICTIONARY_ORDERED 1
-#define ARROW_FLAG_NULLABLE 2
-#define ARROW_FLAG_MAP_KEYS_SORTED 4
-
-struct ArrowSchema {
-  // Array type description
-  const char* format;
-  const char* name;
-  const char* metadata;
-  int64_t flags;
-  int64_t n_children;
-  struct ArrowSchema** children;
-  struct ArrowSchema* dictionary;
-
-  // Release callback
-  void (*release)(struct ArrowSchema*);
-  // Opaque producer-specific data
-  void* private_data;
-};
-
-struct ArrowArray {
-  // Array data description
-  int64_t length;
-  int64_t null_count;
-  int64_t offset;
-  int64_t n_buffers;
-  int64_t n_children;
-  const void** buffers;
-  struct ArrowArray** children;
-  struct ArrowArray* dictionary;
-
-  // Release callback
-  void (*release)(struct ArrowArray*);
-  // Opaque producer-specific data
-  void* private_data;
-};
-
-#endif  // ARROW_C_DATA_INTERFACE
-
-#ifndef ARROW_C_STREAM_INTERFACE
-#define ARROW_C_STREAM_INTERFACE
-
-struct ArrowArrayStream {
-  // Callback to get the stream type
-  // (will be the same for all arrays in the stream).
-  //
-  // Return value: 0 if successful, an `errno`-compatible error code otherwise.
-  //
-  // If successful, the ArrowSchema must be released independently from the stream.
-  int (*get_schema)(struct ArrowArrayStream*, struct ArrowSchema* out);
-
-  // Callback to get the next array
-  // (if no error and the array is released, the stream has ended)
-  //
-  // Return value: 0 if successful, an `errno`-compatible error code otherwise.
-  //
-  // If successful, the ArrowArray must be released independently from the stream.
-  int (*get_next)(struct ArrowArrayStream*, struct ArrowArray* out);
-
-  // Callback to get optional detailed error information.
-  // This must only be called if the last stream operation failed
-  // with a non-0 return code.
-  //
-  // Return value: pointer to a null-terminated character array describing
-  // the last error, or NULL if no description is available.
-  //
-  // The returned pointer is only valid until the next operation on this stream
-  // (including release).
-  const char* (*get_last_error)(struct ArrowArrayStream*);
-
-  // Release callback: release the stream's own resources.
-  // Note that arrays returned by `get_next` must be individually released.
-  void (*release)(struct ArrowArrayStream*);
-
-  // Opaque producer-specific data
-  void* private_data;
-};
-
-#endif  // ARROW_C_STREAM_INTERFACE
-#endif  // ARROW_FLAG_DICTIONARY_ORDERED
 
 /// \file Arrow C Implementation
 ///
@@ -142,26 +57,6 @@ void* ArrowRealloc(void* ptr, int64_t size);
 /// \brief Free a pointer allocated using ArrowMalloc() or ArrowRealloc().
 void ArrowFree(void* ptr);
 
-/// \brief Array buffer allocation and deallocation
-///
-/// Container for allocate, reallocate, and free methods that can be used
-/// to customize allocation and deallocation of buffers when constructing
-/// an ArrowArray.
-struct ArrowBufferAllocator {
-  /// \brief Allocate a buffer or return NULL if it cannot be allocated
-  uint8_t* (*allocate)(struct ArrowBufferAllocator* allocator, int64_t size);
-
-  /// \brief Reallocate a buffer or return NULL if it cannot be reallocated
-  uint8_t* (*reallocate)(struct ArrowBufferAllocator* allocator, uint8_t* ptr,
-                         int64_t old_size, int64_t new_size);
-
-  /// \brief Deallocate a buffer allocated by this allocator
-  void (*free)(struct ArrowBufferAllocator* allocator, uint8_t* ptr, int64_t size);
-
-  /// \brief Opaque data specific to the allocator
-  void* private_data;
-};
-
 /// \brief Return the default allocator
 ///
 /// The default allocator uses ArrowMalloc(), ArrowRealloc(), and
@@ -181,12 +76,6 @@ struct ArrowBufferAllocator* ArrowBufferAllocatorDefault();
 struct ArrowError {
   char message[1024];
 };
-
-/// \brief Return code for success.
-#define NANOARROW_OK 0
-
-/// \brief Represents an errno-compatible error code
-typedef int ArrowErrorCode;
 
 /// \brief Set the contents of an error using printf syntax
 ArrowErrorCode ArrowErrorSet(struct ArrowError* error, const char* fmt, ...);
@@ -483,24 +372,7 @@ ArrowErrorCode ArrowSchemaViewInit(struct ArrowSchemaView* schema_view,
 
 /// }@
 
-/// \defgroup nanoarrow-buffer-builder Growable buffer builders
-
-/// \brief An owning mutable view of a buffer
-struct ArrowBuffer {
-  /// \brief A pointer to the start of the buffer
-  ///
-  /// If capacity_bytes is 0, this value may be NULL.
-  uint8_t* data;
-
-  /// \brief The size of the buffer in bytes
-  int64_t size_bytes;
-
-  /// \brief The capacity of the buffer in bytes
-  int64_t capacity_bytes;
-
-  /// \brief The allocator that will be used to reallocate and/or free the buffer
-  struct ArrowBufferAllocator* allocator;
-};
+/// \defgroup nanoarrow-buffer Owning, growable buffers
 
 /// \brief Initialize an ArrowBuffer
 ///
@@ -594,13 +466,16 @@ struct ArrowBitmapBuilder {
 void ArrowBitmapBuilderInit(struct ArrowBitmapBuilder* bitmap_builder);
 
 /// \brief Append a boolean value to a bitmap
-ArrowErrorCode ArrowBitmapBuilderAppend(struct ArrowBitmapBuilder* bitmap_builder, int8_t value);
+ArrowErrorCode ArrowBitmapBuilderAppend(struct ArrowBitmapBuilder* bitmap_builder,
+                                        int8_t value);
 
 /// \brief Append boolean values encoded as int8_t to a bitmap
-ArrowErrorCode ArrowBitmapBuilderAppendInt8(struct ArrowBitmapBuilder* bitmap_builder, const int8_t* values, int64_t n_values);
+ArrowErrorCode ArrowBitmapBuilderAppendInt8(struct ArrowBitmapBuilder* bitmap_builder,
+                                            const int8_t* values, int64_t n_values);
 
 /// \brief Append boolean values encoded as int32_t to a bitmap
-ArrowErrorCode ArrowBitmapBuilderAppendInt32(struct ArrowBitmapBuilder* bitmap_builder, const int32_t* values, int64_t n_values);
+ArrowErrorCode ArrowBitmapBuilderAppendInt32(struct ArrowBitmapBuilder* bitmap_builder,
+                                             const int32_t* values, int64_t n_values);
 
 /// \brief Force pending value to be written to buffer
 ///
